requirements.txt
# Dinner Party Utility Requirements
# Python 3.7+ required for type hints and other features

# Unit conversion library
pint>=0.20.0

# OpenAI API for natural language parsing
openai>=1.0.0

# Pydantic for structured output schemas
pydantic>=2.0.0

# YAML parsing for config files
pyyaml>=6.0
